Add "system" certificate configuration
This new config loads `/etc/ssl/certs/ca-certificates.crt` on Ubuntu and
`/etc/ssl/cert.pem` on OSX, which are the system-managed certificate
stores on the respective platforms.

As the paths are now absolute, a dependent application will no longer
be tightly coupled to the build path. This fixes outgoing HTTPS calls
on Heroku (verified manually with a test project).