Replace #if os() test with different well-known paths and fallback
With this change, the `system` configuration actively searches for the
cert files, ultimately returning nil if none of them was found (which
causes the openBSD certs to be loaded, as per the previous defaults).

The list is somewhat futureproof, should Swift starts to support other
Linux distributions. In the meanwhile, no application will break.